+++ front/components/assistant/conversation/SpaceAboutTab.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export function SpaceAboutTab() {
+  return (
+    <div className="flex w-full items-center justify-center p-8">
+      <div className="text-center text-muted-foreground dark:text-muted-foreground-night">
+        About this project tab - Coming soon
+      </div>
+    </div>
+  );
+}
